Switched from Vuforia to TensorFlow. Vuforia code was moved to the archived code folder. TensorFlow code is the sample from the FTC SDK v5.2 (I think).
